requirements.txt
accelerate==0.23.0
biopython==1.78
cloudpathlib==0.15.0
coolname==2.2.0
crcmod==1.7
datasets==2.14.5
devtools==0.11.0
faiss-cpu==1.7.4
fire==0.5.0
gcsfs==2023.6.0
hnswlib==0.7.0
imagecodecs==2023.9.18
imageio==2.26.0
matplotlib==3.7.1
monai==1.3.0
more_itertools==9.1.0
numba==0.57.0
numpy==1.24.3
pandas==1.5.3
pydantic-settings==2.0.3
pydantic==2.4.2
pytest-xdist==3.3.1
pytest==7.4.2
python-dotenv==1.0.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
statsmodels==0.13.5
torch==2.0.1
tqdm==4.65.0
transformers[torch]==4.34.1
wandb==0.15.12